Move the `HwBench` struct to the sysinfo crate
