+++ client/sysinfo/Cargo.toml
@@ -20,5 +20,7 @@ rand = "0.7.3"
 rand_pcg = "0.2.1"
 regex = "1"
 libc = "0.2"
+serde = { version = "1.0.136", features = ["derive"] }
+serde_json = "1.0.79"
 sp-core = { version = "6.0.0", path = "../../primitives/core" }
 sc-telemetry = { version = "4.0.0-dev", path = "../telemetry" }
